\name{CIplot-package}
\alias{CIplot-package}
\docType{package}
\title{
\packageTitle{CIplot}
}
\description{
\packageDescription{CIplot}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CIplot}
\packageIndices{CIplot}
}
\author{
\packageAuthor{CIplot}

Maintainer: \packageMaintainer{CIplot}
}
%% \references{
%% %% ~~ Literature or other references for background information ~~
%% }
%% ~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
%% ~~ the R documentation directory ~~
\keyword{ package }
%% \seealso{
%% %% ~~ Optional links to other man pages, e.g. ~~
%% %% ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%% }
%% \examples{
%% %% ~~ simple examples of the most important functions ~~
%% }
