% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ORci.R
\name{print.ORci}
\alias{print.ORci}
\title{Print Methods for Odds Ratios and their Confidence Intervals
of \code{ORci} object}
\usage{
\method{print}{ORci}(x, ...)
}
\arguments{
\item{x}{\code{ORci} object.see alse \code{\link{ORci}}.}

\item{\dots}{other options for print such as \code{digits}.}
}
\description{
Print odds ratios and their confidence intervals of \code{ORci} object.
}
\examples{
require(MASS)
data(birthwt)
x <- glm(low ~  age + lwt + smoke + ptl + ht + ui, data = birthwt,
         family = binomial)
OR1 <- ORci(x)
print(OR1, digits = 3)

}
\seealso{
\code{glm}, \code{\link{ORci}}.
}
