\name{CIplot-package}
\alias{CIplot-package}
\docType{package}
\title{
\packageTitle{CIplot}
}
\description{
\packageDescription{CIplot}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CIplot}
\packageIndices{CIplot}
}
\author{
\packageAuthor{CIplot}

Maintainer: \packageMaintainer{CIplot}
}
\keyword{ package }
